body: Start work on the new cpBody wrapper
Currently this is a wrapper around cpBody and in a single module, but we'll
split it up and make it nice. I know others might object to our use of toplevel
classes and not struct casts, however we'd need to rely on ugly cpBodyNew
semantics which will cause allocation anyway.

Instead go with the high level approach so in future we can have signals and
whatnot on bodies.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>