Convert everything to fat pointers
Unfortunately mixing D and C memory when it comes to chipmunkd is going to
be a collosal pain in the arse, so instead we're going to stuff the pointers
into our objects and let C manage the memory.

We've still got some dodgy crashes and undefined behaviour but we'll get
there.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>