body: Maintain an orphan list of pending shapes
We can't get a list of shapes from the space because they were never
added to the body, as only the space can add them directly to the
body.

To counter this chicken and egg situation, whilst maintaining a 'nice'
API, we maintain an orphan list of pending shape additions, and when
we get a new world, we register all of the shapes.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>